[extensiontest] Remove StatusWatcher helpers
