+++ .chloggen/mx-psi_move-extensionstatus.yaml
@@ -1,0 +1,26 @@
+# Use this changelog template to create an entry for release notes.
+
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: breaking
+
+# The name of the component, or a single word describing the area of concern, (e.g. otlpreceiver)
+component: Remove `extensiontest` StatusWatcher helpers
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: | 
+  They were unused. They may be added back on a different module or after `componentstatus` is marked 1.0
+
+# One or more tracking issues or pull requests related to the change
+issues: [11044]
+
+# (Optional) One or more lines of additional information to render under the primary note.
+# These lines will be padded with 2 spaces and then inserted directly into the document.
+# Use pipe (|) for multiline entries.
+subtext:
+
+# Optional: The change log or logs in which this entry should be included.
+# e.g. '[user]' or '[user, api]'
+# Include 'user' if the change is relevant to end users.
+# Include 'api' if there is a change to a library API.
+# Default: '[user]'
+change_logs: [api]
