+++ extension/auth/go.mod
@@ -59,5 +59,3 @@ replace go.opentelemetry.io/collector/extension => ../
 replace go.opentelemetry.io/collector/pdata => ../../pdata
 
 replace go.opentelemetry.io/collector/config/configtelemetry => ../../config/configtelemetry
-
-replace go.opentelemetry.io/collector/component/componentstatus => ../../component/componentstatus
