--- extension/extensiontest/statuswatcher_extension.go
@@ -1,48 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package extensiontest // import "go.opentelemetry.io/collector/extension/extensiontest"
-
-import (
-	"context"
-
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/component/componentstatus"
-	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/extension"
-)
-
-// NewStatusWatcherExtensionCreateSettings returns a new nop settings for Create*Extension functions.
-func NewStatusWatcherExtensionCreateSettings() extension.Settings {
-	return extension.Settings{
-		TelemetrySettings: componenttest.NewNopTelemetrySettings(),
-		BuildInfo:         component.NewDefaultBuildInfo(),
-	}
-}
-
-// NewStatusWatcherExtensionFactory returns a component.ExtensionFactory to construct a status watcher extension.
-func NewStatusWatcherExtensionFactory(
-	onStatusChanged func(source *componentstatus.InstanceID, event *componentstatus.Event),
-) extension.Factory {
-	return extension.NewFactory(
-		component.MustNewType("statuswatcher"),
-		func() component.Config {
-			return &struct{}{}
-		},
-		func(context.Context, extension.Settings, component.Config) (component.Component, error) {
-			return &statusWatcherExtension{onStatusChanged: onStatusChanged}, nil
-		},
-		component.StabilityLevelStable)
-}
-
-// statusWatcherExtension receives status events reported via component status reporting for testing
-// purposes.
-type statusWatcherExtension struct {
-	component.StartFunc
-	component.ShutdownFunc
-	onStatusChanged func(source *componentstatus.InstanceID, event *componentstatus.Event)
-}
-
-func (e statusWatcherExtension) ComponentStatusChanged(source *componentstatus.InstanceID, event *componentstatus.Event) {
-	e.onStatusChanged(source, event)
-}
